requirements.txt
tensorflow==2.10
pillow==10.4.0
python-multipart==0.0.9
